+++ src/libs/shared/src/canister.rs
@@ -23,7 +23,7 @@ use ic_cdk::api::stable::{stable_size, WASM_PAGE_SIZE_IN_BYTES};
 ///
 pub fn memory_size() -> MemorySize {
     MemorySize {
-        heap: wasm_memory_size(0) * WASM_PAGE_SIZE_IN_BYTES as usize,
-        stable: stable_size() as usize * WASM_PAGE_SIZE_IN_BYTES as usize,
+        heap: wasm_memory_size(0) as u64 * WASM_PAGE_SIZE_IN_BYTES,
+        stable: stable_size() * WASM_PAGE_SIZE_IN_BYTES,
     }
 }
